+++ src/main/java/com/techelevator/Candy.java
@@ -1,10 +1,9 @@
 package com.techelevator;
 
-import java.math.BigDecimal;
 
 public class Candy extends Items {
 	
-	public Candy(String name ,BigDecimal price, String code, double quantity) {
+	public Candy(String name ,double price, String code, double quantity) {
 		super(name, price, code, quantity);
 	}
